perf(es5): remove legacy es5 bundling resources #20
